Fix prompt path display
This commit makes two changes. First, it replaces %3C with %3c, so that
the value of $HOME gets replaced with ~ in the prompt. Second, it
increases %3(c||) to %4(c||), so that the conditional is only true when
we are 4 directories deep; this ensures that the ellipsis is only
prepended when we go beyond 3 directories deep, i.e. it won't ever show
…/~/foo/bar.