Fix man page width
The man command formats its output according to the current terminal
width. This is independent of the pager used; it is done before the
output is passed to the terminal, thus preventing pagers from reflowing
the text upon SIGWINCH (screen resize). Therefore we fix the man page
width to a readable 80 columns, which should also fit in most
vertically-split windows on modern machines.