Consider unnamed buffers to be inside cwd
Previously, a buffer without a name — such as a buffer created by :enew,
or a buffer created by using Vim as Git pager — would display our
not-in-cwd icon in the statusline. This commit changes this behaviour so
that an unnamed buffer does not show this icon, since the semantic of
"this file is not in the current working directory" doesn't make much
sense if "this file" is effectively null.