Allow gf to find things relative to the repo root
The gf command opens the file whose name is under the cursor. With
monorepos, this can get annoying when our cwd is a subdirectory of the
repo root, since gf will by default look for files under the cwd. This
commit allows Vim to look up the Git (or Hg!) repo root, and find things
relative to that path.