Teach packages how to install themselves
This commit moves the knowledge of which particular flags Stow should
use for each package into that package. The install script just invokes
per-package install scripts.

This also allows easily invoking the install script for a particular
package. The package install scripts are written to always execute from
the directory they're in, regardless of where we're invoking them from.