Work around Stow not supporting --dotfiles on dirs
The old .vim/ location was problematic due to an unexpected three-way
conflict.

1. We use .vim/ relative to a repo root to store the Vim session file.
2. The canonical path where Vim stores its configs relative to ~/ is
   .vim/.
3. Stow does not currently support the --dotfiles flag properly for
   directories, so our Vim config relative to the vim/ package must be
   named .vim/ (unless we intervene; read on).

Due to #1, we had .vim/ in our gitignore, so that Vim session files
wouldn't show up to Git operations. However, due to #2, this meant that
vim/.vim/ in this repo was not being tracked properly by Git. We could
rename vim/.vim/ in order to work around this, if not for #3.

This commit works around this issue by renaming vim/.vim/ to
vim/dot-vim/, excluding vim/dot-vim/ from the Stow invocation for vim/,
and instead setting up vim/dot-vim/ as a Stow package in its own right
that gets installed to ~/.vim/.