Collapse Git root into branch name if they match
This commit allows the prompt pwd display to "subsume" the Git branch
display, if we are inside a Git repo whose root directory name matches
the name of the current branch (allowing for transforming between / and
--, as is my convention for Git worktrees). By "subsume", I mean: the
Git branch is no longer displayed as a separate prompt element; and the
part of the pwd that matches the branch name is coloured green, like the
branch name normally is, instead of yellow, like the pwd normally is.

This commit also adds an indicator for when we are on a detached HEAD.

This commit also gets rid of some unnecessary prompt escape sequences.