Improve handling of "modifier" abbreviations
This changes the way modifiers like sudo and watch are handled. Now, we
create two versions of each normal abbreviation: a regular version, and
a global version prepended with an @. Modifier abbreviations expand to
the modifier followed by a space and an @, e.g. "s" becomes "sudo @".
This then allows a sequence such as "s d" to expand first to "sudo @",
then add the "d", then expand to "sudo docker".