Update fzf.vim
This commit addresses a few small issues.

First, it adds an explicit note about the dependency on a particular
version of the fzf binary, which will be useful in contexts where fzf
can't easily be kept up to date.

Second, it makes :Rg output more readable.

Third, it removes an argument to :Rg that was deprecated in
fd812b57 (2023-06-04). This is not a functional change; the extra
argument was silently skipped.

Fourth, it slightly reduces the size of the :Rg preview pane in order to
fit more search results on the screen.