Fix compinit sequence
This change ensures that we only call compinit once, and that things are
ordered sensibly. Delegating compinit to zcomet might mean we no longer
have the benefit of our old once-daily-cache-invalidation optimisation.
However, we do now see an improvement of ~79 ms on Zsh load time
compared to the unoptimised version with repeated compinit calls.