Automatically pass env hash to tools
This commit changes .detect_env such that it no longer exports many
variables, but rather a single associative array. However, Zsh does not
actually allow exporting an associative array to child processes. So we
implement a magic system that allows tools to automagically "pull in"
the associative array of env data if they need it.